Adds application-specific evaluator example
This adds an application-specific  `OceanEvaluator` to demonstrate
evaluation driven from error-analysis.

While this is extremely simple, this helps us from accidentally teaching
people that picking pre-built metrics is the only path to Q/A, or even
a relevant one.

Here's a relevant video on common eval mistakes (over-reliance on
built-in evals like hallucination) by Shreya Shankar and Hamel Husain:

https://www.youtube.com/watch?v=GL0XhAj5LPE&ab_channel=HamelHusain

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>